feat(game-vault): Implementação de login por autenticação via token

+++ controllers/users_controller.ts
@@ -1,10 +1,16 @@
 
 import type { HttpContext } from '@adonisjs/core/http'
 import UsersService from '#services/users_service'
-import User from '#models/user'
 
 
 export default class UsersController {
+  async login({ request }: HttpContext) {
+    const { email, password } = request.all()
+    if(!email || !password) throw new Error('Invalid credentials')
+
+    return new UsersService().login(email, password)
+  }
+
  async show({ params }: HttpContext) {
         const id = params.id
         return new UsersService().show(id)
@@ -24,8 +30,8 @@ export default class UsersController {
     }
 
     async delete({ params }: HttpContext) {
-        const user = await User.findOrFail(params.id)
-        await user.delete()
+        const id = params.id
+        if(!id) throw new Error('User not found')
 
         return new UsersService().delete(params.id)
     }
